Changes naming convention of Docker to openzipkin-contrib
Docker Hub registry doesn't like hyphens, but GitHub is our default
registry, and the org has to match the first part of the registry path.

This renames the docker images from openzipkincontrib/X to
openzipkin-contrib/X

While this is unfortunate, I think aligning names is better use of time
and attention in contrib vs publishing to two registries and have the
names conflict. Especially, as in the future ideally this will become a
non-contrib project anyway.